Add model files, add test build for testpypi, remove old prime and primesh modules, rename main module to prime from prime_torch, small bugfixes, remove documentation typos
